Testing support for push notifications
Added a controller for push notifications with a function to store the
user subscription to the database, and a show function with temporary
data to test push notifications.

Added a Notification class which sends a temporary push notification to
subscribed clients.

Also added API routes for subscribing to push notifications, and to
schedule a notification.